refactor(ai): expose side-effect-free API registry
