+++ packages/ai/src/compat.ts
@@ -50,12 +50,7 @@ import { builtinModels, getBuiltinModel, getBuiltinModels, getBuiltinProviders }
 
 export type { BuiltinProvider } from "./providers/all.ts";
 
-import {
-	createFauxCore,
-	type FauxProviderRegistration,
-	getRegisteredFauxProvider,
-	type RegisterFauxProviderOptions,
-} from "./providers/faux.ts";
+import { createFauxCore, type FauxProviderRegistration, type RegisterFauxProviderOptions } from "./providers/faux.ts";
 import {
 	getProtocol,
 	getToolCallFormat,
@@ -72,7 +67,6 @@ import type {
 	ProviderStreamOptions,
 	ProviderStreams,
 	SimpleStreamOptions,
-	StreamFunction,
 	StreamOptions,
 } from "./types.ts";
 
@@ -85,102 +79,18 @@ export const getModels = getBuiltinModels;
 /** @deprecated Static catalog read. Use `getBuiltinProviders` from "@earendil-works/pi-ai/providers/all" or `Models.getProviders()`. */
 export const getProviders = getBuiltinProviders;
 
-export type ApiStreamFunction = (
-	model: Model<Api>,
-	context: Context,
-	options?: StreamOptions,
-) => AssistantMessageEventStream;
+export {
+	type ApiProvider,
+	type ApiProviderInternal,
+	type ApiStreamFunction,
+	type ApiStreamSimpleFunction,
+	getApiProvider,
+	getApiProviders,
+	registerApiProvider,
+	unregisterApiProviders,
+} from "./api-registry.ts";
 
-export type ApiStreamSimpleFunction = (
-	model: Model<Api>,
-	context: Context,
-	options?: SimpleStreamOptions,
-) => AssistantMessageEventStream;
-
-export interface ApiProvider<TApi extends Api = Api, TOptions extends StreamOptions = StreamOptions> {
-	api: TApi;
-	stream: StreamFunction<TApi, TOptions>;
-	streamSimple: StreamFunction<TApi, SimpleStreamOptions>;
-}
-
-interface ApiProviderInternal {
-	api: Api;
-	stream: ApiStreamFunction;
-	streamSimple: ApiStreamSimpleFunction;
-}
-
-type RegisteredApiProvider = {
-	provider: ApiProviderInternal;
-	sourceId?: string;
-};
-
-const apiProviderRegistry = new Map<string, RegisteredApiProvider>();
-
-function wrapStream<TApi extends Api, TOptions extends StreamOptions>(
-	api: TApi,
-	stream: StreamFunction<TApi, TOptions>,
-): ApiStreamFunction {
-	return (model, context, options) => {
-		if (model.api !== api) {
-			throw new Error(`Mismatched api: ${model.api} expected ${api}`);
-		}
-		return stream(model as Model<TApi>, context, options as TOptions);
-	};
-}
-
-function wrapStreamSimple<TApi extends Api>(
-	api: TApi,
-	streamSimple: StreamFunction<TApi, SimpleStreamOptions>,
-): ApiStreamSimpleFunction {
-	return (model, context, options) => {
-		if (model.api !== api) {
-			throw new Error(`Mismatched api: ${model.api} expected ${api}`);
-		}
-		return streamSimple(model as Model<TApi>, context, options);
-	};
-}
-
-export function registerApiProvider<TApi extends Api, TOptions extends StreamOptions>(
-	provider: ApiProvider<TApi, TOptions>,
-	sourceId?: string,
-): void {
-	apiProviderRegistry.set(provider.api, {
-		provider: {
-			api: provider.api,
-			stream: wrapStream(provider.api, provider.stream),
-			streamSimple: wrapStreamSimple(provider.api, provider.streamSimple),
-		},
-		sourceId,
-	});
-}
-
-export function getApiProvider(api: Api): ApiProviderInternal | undefined {
-	const faux = getRegisteredFauxProvider(api);
-	if (faux) {
-		return {
-			api: faux.api,
-			stream: wrapStream(faux.api, faux.stream),
-			streamSimple: wrapStreamSimple(faux.api, faux.streamSimple),
-		};
-	}
-	return apiProviderRegistry.get(api)?.provider;
-}
-
-export function getApiProviders(): ApiProviderInternal[] {
-	return Array.from(apiProviderRegistry.values(), (entry) => entry.provider);
-}
-
-export function unregisterApiProviders(sourceId: string): void {
-	for (const [api, entry] of apiProviderRegistry.entries()) {
-		if (entry.sourceId === sourceId) {
-			apiProviderRegistry.delete(api);
-		}
-	}
-}
-
-function clearApiProviders(): void {
-	apiProviderRegistry.clear();
-}
+import { clearApiProviders, getApiProvider, registerApiProvider, unregisterApiProviders } from "./api-registry.ts";
 
 export function registerFauxProvider(options: RegisterFauxProviderOptions = {}): FauxProviderRegistration {
 	const core = createFauxCore(options);
